Funciones para eliminar zombies (falta)

+++ ArchivoCabecera(.h)/Disparo.h
@@ -9,7 +9,7 @@ class DISPARO
 private: 
 
 	int TIEMPODISPARO;
-	int VELOCIDAD;
+	float VELOCIDAD;
 
 	char DIRECCION;
